suppress redirect when not on teams view
The SelectTaskInSupervisor Action navigates to the teams view if that's not the active view already. We do this when `teamViewTaskSID` has been set in localStorage, but that value is never unset. So subsequent page loads in Flex UI will always default the view back to the Teams View. We should only trigger the action and stage the monitor config if we're running a page refresh.